Revert "Cleaned code, moved verifying clients"
This reverts commit 9249c37517d7884ff86d598435e3f86f6b2d0375.
