+++ server/src/client/index.ts
@@ -105,6 +105,7 @@ export default class WebSocketClient {
   }
 
   async interval(): Promise<void> {
+    this.ws.send(JSON.stringify(this.getDevicesStatus()));
     this.getCurrentConnectionWithAccess().forEach(deviceObject => {
       const deviceData: DeviceDataClient = {
         deviceUid: deviceObject.deviceUid,
