+++ server/src/services/websocket_devices.ts
@@ -1,13 +1,8 @@
 import { IncomingMessage } from 'http';
 import WebSocket from 'ws';
-import {
-  logSocketConnection,
-  logSocketAttempt,
-  logMissing,
-  logInvalid,
-} from '@/cli';
-import jwt from 'jsonwebtoken';
+import { logSocketConnection } from '@/cli';
 import chalk from 'chalk';
+import { verifyDevice } from '@/auth';
 import http from 'http';
 import { DeviceType, DevicesTypes } from '@gbaranski/types';
 import { logError } from '@/cli';
@@ -15,42 +10,8 @@ import WatermixerDevice from '@/devices/watermixer';
 import Device, { AnyDeviceObject } from '@/devices';
 import AlarmclockDevice from '@/devices/alarmclock';
 import { validateDevice } from './firebase';
-import { VerifyInfo, VerifyCallback } from '@/auth';
 
 const httpServer = http.createServer();
-
-export const verifyDevice = (
-  info: VerifyInfo,
-  callback: VerifyCallback,
-): void => {
-  logSocketAttempt(
-    info.req,
-    info.req.headers['devicetype'] || 'unknown',
-    'device',
-  );
-  if (!process.env.JWT_KEY) throw new Error('Missing process.env.JWT_KEY');
-  const token = info.req.headers.token || '';
-  if (!token) {
-    logMissing('JWT token');
-    callback(false, 401, 'Unauthorized');
-    return;
-  }
-
-  jwt.verify(token as string, process.env.JWT_KEY, (err, decoded) => {
-    if (!decoded) {
-      logMissing('decoded username at JWT Token');
-      callback(false, 400, 'Missing decoded username');
-      return;
-    }
-    if (err) {
-      callback(false, 401, 'Unauthorized');
-      logInvalid('token');
-    } else {
-      info.req.headers.device = (decoded as { device: string }).device;
-      callback(true);
-    }
-  });
-};
 
 export const wss: WebSocket.Server = new WebSocket.Server({
   server: httpServer,
